+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Released 2016-mm-dd
 
+ - Add method to ignore param for the node id generation
  - Replace `validate` with `beforeCreate` option to modify nodes before create/return them
  - Option to bump version an force analysis recalculation.
  - Optional params are casted to null when not provided.